requirements.txt
youtube_transcript_api>=0.6.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiohttp>=3.9.0 